fix(2fa): extract TwoFactorChallenge into its own component file to satisfy Next.js page export rules
