+++ frontend/tests/components/TwoFactor.test.jsx
@@ -2,7 +2,7 @@
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import TwoFactorSetup from '../../app/settings/security/page.jsx';
-import TwoFactorChallenge from '../../app/auth/2fa/page.jsx';
+import TwoFactorChallenge from '../../app/auth/2fa/TwoFactorChallenge.jsx';
 
 // Mock fetch globally
 global.fetch = jest.fn();
